2025/solution/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langfuse>=2.0.0
ulid-py>=1.1.0
langgraph>=0.2.0
langchain-core>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
geopy>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
